refactor(waiting): 스키마 이름 변경
