+++ src/v1/sockets/waiting/schemas/auto-join.schema.ts
@@ -1,8 +1,8 @@
 import { TypeOf, z } from 'zod';
 import { tournamentSizeSchema } from './tournament.schema.js';
 
-export type customJoinSchemaType = TypeOf<typeof customJoinSchema>;
+export type customJoinSchemaType = TypeOf<typeof autoJoinSchema>;
 
-export const customJoinSchema = z.object({
+export const autoJoinSchema = z.object({
   tournamentSize: tournamentSizeSchema,
 });
